\name{APO1-package}
\alias{APO1-package}
\alias{APO1}
\docType{package}
\title{
\packageTitle{APO1}
}
\description{
\packageDescription{APO1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{APO1}
\packageIndices{APO1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{APO1}

Maintainer: \packageMaintainer{APO1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
